requirements.txt
# FX3110 Monitor - Python Dependencies
# Minimum Python version: 3.8
python-dotenv==1.0.0
pyyaml>=6.0
